// types/typeChauffageActuel.ts
/**
 * Enum for current heating system types
 * Used in step 2 (Chauffage Actuel)
 */
export enum TypeChauffageActuel {
  FIOUL = "Fioul",
  GAZ = "Gaz",
  GPL = "GPL",
  PELLETS = "Pellets",
  BOIS = "Bois",
  ELECTRIQUE = "Electrique",
  PAC_AIR_AIR = "PAC Air/Air",
  PAC_AIR_EAU = "PAC Air/Eau",
  PAC_EAU_EAU = "PAC Eau/Eau",
}